changed bytes to float

+++ src/RayTracing/Light.java
@@ -2,9 +2,9 @@ package RayTracing;
 
 public class Light {
     public Vector Position;
-    public byte RedColor;
-    public byte GreenColor;
-    public byte BlueColor;
+    public double RedColor;
+    public double GreenColor;
+    public double BlueColor;
     public double SpecularIntensity;
     public double ShadowIntensity;
     public double LightRadius;
